refactor: _entities와 _utilities로 디렉토리 이름 변경
모든 곳에서 공통적으로 사용되는 디렉토리를 src 디렉토리 상단에 두기 위해 디렉토리명 수정
