+++ src/_entities/FeedImgs.ts
