+++ src/_entities/HashTags.ts
