+++ src/_entities/Persona.ts
