+++ src/_entities/ProfileHashTagMapping.ts
