+++ src/_entities/QuestionContent.ts
